Moved load_qwen_vision_addon and QwenVisionComponents to vision add-on file
